pyproject.toml
[project]
name = "screenshot-tool"
version = "1.0.0"
description = "Automated multi-language GUI screenshot capture tool"
requires-python = ">=3.10"
dependencies = [
    "pywin32>=305",
    "pillow>=10.0.0",
    "psutil>=5.9.0",
    "pyautogui>=0.9.54",
    "keyboard>=0.13.5",
    "pywinauto>=0.6.8",
]

[project.scripts]
screenshot-tool = "screenshot_tool.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
